Moved GetPreviewBitmap to SolidWorksApplication
