+++ SolidDna/CADBooster.SolidDna/SolidWorks/Application/Interop/IPictureDisp.cs
